requirements.txt
bottle==0.12.8
gnureadline==6.3.3
ipython==3.2.1
mecab-python3==0.7
numpy==1.9.2
pandas==0.16.2
pep8==1.6.2
pyflakes==0.9.2
python-dateutil==2.4.2
pytz==2015.4
scikit-learn==0.16.1
scipy==0.15.1
six==1.9.0
virtualenv==13.1.0
